03_user_management_system/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
